mass code cleanup, clustering done in eta/phi

+++ scripts_condor/normalize.sh
@@ -37,57 +37,59 @@ then
 	#hadd all the jobs for this sample
 	echo "/mnt/hadoop/${inputDir}/${sample}*_Job*.root"
 	hadd ${sample}.root /mnt/hadoop/${inputDir}/${sample}*_Job*.root
+	output=${sample}.root
+	if [ ${isData} == "no" ]
+        	then
+		eval `scramv1 runtime -sh`
+		if [ -f $CMSSW_BASE/src/llp_analyzer/data/xSections.dat ]
+		then
+			mkdir -p data
+			cp $CMSSW_BASE/src/llp_analyzer/data/xSections.dat data/xSections.dat 
+		else
+			echo "data/xSections.dat doesn't exist"
 
+		fi	
+		
+		#create normalization file
+		rm -f $normalize_file
+		if [ "$mode" == "bkg" ]
+		then
+			dataset=${sample}
+		else
+			dataset=${sample/*_vh_/}
+        		dataset=${dataset/_mx*/}
+        		dataset=${mode}_${dataset}
+		fi
 
-	eval `scramv1 runtime -sh`
-	if [ -f $CMSSW_BASE/src/llp_analyzer/data/xSections.dat ]
-	then
-		mkdir -p data
-		cp $CMSSW_BASE/src/llp_analyzer/data/xSections.dat data/xSections.dat 
-	else
-		echo "data/xSections.dat doesn't exist"
+		echo "$dataset"
+		echo "${dataset} ${runDir}/${sample}.root" > $normalize_file
 
-	fi	
-	
-	#create normalization file
-	rm -f $normalize_file
-	if [ "$mode" == "bkg" ]
-	then
-		dataset=${sample}
-	else
-		dataset=${sample/*_vh_/}
-        	dataset=${dataset/_mx*/}
-        	dataset=${mode}_${dataset}
+		if [ -f $normalize_file ]
+		then
+			echo "normalization file created"
+		fi
+
+		#normalize
+		if [ -f $CMSSW_BASE/src/llp_analyzer/NormalizeNtuple ]
+        	then
+        	        cp $CMSSW_BASE/src/llp_analyzer/NormalizeNtuple ./
+		        ./NormalizeNtuple ${normalize_file} 1
+		else
+			echo "NormalizeNtuple not found"
+		fi
+		echo "Normalization done"
+		output=${sample}_1pb_weighted.root
 	fi
-
-	echo "$dataset"
-	echo "${dataset} ${runDir}/${sample}.root" > $normalize_file
-
-	if [ -f $normalize_file ]
-	then
-		echo "normalization file created"
-	fi
-
-	#normalize
-	if [ -f $CMSSW_BASE/src/llp_analyzer/NormalizeNtuple ]
-        then
-                cp $CMSSW_BASE/src/llp_analyzer/NormalizeNtuple ./
-	        ./NormalizeNtuple ${normalize_file} 1
-	else
-		echo "NormalizeNtuple not found"
-	fi
-	echo "Normalization done"
-	
 	sleep 2
         echo "I slept for 2 second"
 	
 	# copy normalized file back to hadoop
 	eval `scram unsetenv -sh`
         gfal-mkdir -p gsiftp://transfer.ultralight.org//${outputDir}
-	gfal-copy -f --checksum-mode=both ${runDir}/${sample}_1pb_weighted.root gsiftp://transfer.ultralight.org//${outputDir}/${sample}_1pb_weighted.root
+	gfal-copy -f --checksum-mode=both ${runDir}/${output} gsiftp://transfer.ultralight.org//${outputDir}/${output}
 	
 
-	if [ -f /mnt/hadoop/${outputDir}/${sample}_1pb_weighted.root ]
+	if [ -f /mnt/hadoop/${outputDir}/${output} ]
 	then
 	        echo "copied succeed"
 	else
